FIR/UAST: merge multi resolve utils
