+++ uast/uast-kotlin-base/src/org/jetbrains/uast/kotlin/internal/multiResolveUtils.kt
@@ -5,9 +5,11 @@
 
 package org.jetbrains.uast.kotlin.internal
 
+import com.intellij.psi.PsiPolyVariantReference
 import com.intellij.psi.PsiSubstitutor
 import com.intellij.psi.ResolveResult
 import com.intellij.psi.infos.CandidateInfo
+import org.jetbrains.kotlin.psi.KtElement
 import org.jetbrains.kotlin.psi.KtExpression
 import org.jetbrains.uast.kotlin.BaseKotlinUastResolveProviderService
 
@@ -22,3 +24,10 @@ fun getResolveResultVariants(
     return referenceVariants.mapNotNull { CandidateInfo(it, PsiSubstitutor.EMPTY) }.asIterable()
 }
 
+fun KtElement.multiResolveResults(): Sequence<ResolveResult> =
+    references.asSequence().flatMap { ref ->
+        when (ref) {
+            is PsiPolyVariantReference -> ref.multiResolve(false).asSequence()
+            else -> (ref.resolve()?.let { sequenceOf(CandidateInfo(it, PsiSubstitutor.EMPTY)) }).orEmpty()
+        }
+    }
